+++ R/uncons.R
@@ -5,6 +5,12 @@
 #'
 #' @param object Fitted 'cirls' object.
 #'
+#' @details
+#'
+#' ## Note on starting values
+#'
+#' If any starting values were provided to fit the `cirls` object, they are not transferred to the fitting of the unconstrained model.
+#'
 #' @returns A [glm][stats::glm()] object.
 #'
 #' @examples
@@ -28,43 +34,39 @@
 #' @export
 uncons <- function(object){
 
-  #----- Extracting everything: cumbersome
-  # # Extract the model.frame and control for glm
-  # mf <- model.frame(object)
-  # mt <- attr(mf, "terms")
-  # control <- object$control[names(object$control) %in%
-  #     methods::formalArgs(stats::glm.control)]
-  #
-  # # Fit with glm.fit
-  # # NOTE: the starting values and singular.ok are ignored
-  # fit <- glm.fit(x = model.matrix(mf), y = model.response(mf),
-  #   weights = model.weights(mf), offset = model.offset(mf),
-  #   family = object$family, control = control,
-  #   intercept = attr(mt, "intercept") > 0L)
-  #
-  # # Put together as in glm
-  # if (length(model.offset(mf)) && attr(mt, "intercept") > 0L) {
-  #   fit2 <- glm.fit(x = model.matrix(mf)[, "(Intercept)", drop = FALSE],
-  #     y = model.response(mf), mustart = fit$fitted.values,
-  #     weights = model.weights(mf), offset = model.offset(mf),
-  #     family = object$family, control = control, intercept = TRUE)
-  #   if (!fit2$converged)
-  #     warning("fitting to calculate the null deviance did not converge -- increase 'maxit'?")
-  #   fit$null.deviance <- fit2$deviance
-  # }
-  # if (!is.null(object$model)) fit$model <- mf
-  # fit$na.action <- attr(mf, "na.action")
-  # if (!is.null(object$x)) fit$x <- model.matrix(mf)
-  # if (is.null(object$y)) fit$y <- NULL
-  # structure(c(fit, list(call = object$call, formula = object$formula, terms = mt,
-  #   data = object$data, offset = model.offset(mf), control = control,
-  #   method = "glm.fit", contrasts = attr(model.matrix(mf), "contrasts"),
-  #   xlevels = .getXlevels(mt, mf))), class = c(fit$class, c("glm", "lm")))
+  #----- Extract the data and fit the model
 
-  #----- Modifying the call instead: works well with the data environment
+  # Extract the model.frame and control for glm
+  mf <- model.frame(object)
+  mt <- attr(mf, "terms")
+  control <- object$control[names(object$control) %in%
+      methods::formalArgs(stats::glm.control)]
+  x <- model.matrix(mt, mf)
+  y <- model.response(mf)
+  weights <- model.weights(mf)
+  offset <- model.offset(mf)
+
+  # Fit with glm.fit
+  # NOTE: the starting values and singular.ok are ignored
+  fit <- glm.fit(x = x, y = y, weights = weights, offset = offset,
+    family = object$family, control = control,
+    intercept = attr(mt, "intercept") > 0L)
+
+  # Part to compute the null deviance in the case of an offset and no intercept
+  if (length(offset) && attr(mt, "intercept") > 0L) {
+    fit2 <- glm.fit(x = x[, "(Intercept)", drop = FALSE],
+      y = y, mustart = fit$fitted.values,
+      weights = weights, offset = offset,
+      family = object$family, control = control, intercept = TRUE)
+    if (!fit2$converged)
+      warning("fitting to calculate the null deviance did not converge -- increase 'maxit'?")
+    fit$null.deviance <- fit2$deviance
+  }
+
+  #----- Modify the call to match the new model
 
   # Extract call and change the method argument
-  call <- as.list(stats::getCall(object))
+  call <- as.list(object$call)
 
   # Remove arguments related to cirls only from the call
   cirlsargs <- c("method", setdiff(names(call)[-1],
@@ -73,13 +75,26 @@ uncons <- function(object){
 
   # Modify the control list if present
   if ("control" %in% names(call)){
-    control <- as.list(call$control)
-    remargs <- setdiff(names(control)[-1],
+    clcontrol <- as.list(call$control)
+    remargs <- setdiff(names(clcontrol)[-1],
       methods::formalArgs(stats::glm.control))
-    control[remargs] <- NULL
-    call$control <- if (length(control) > 1) as.call(control) else NULL
+    clcontrol[remargs] <- NULL
+    call$control <- if (length(clcontrol) > 1) as.call(clcontrol) else NULL
   }
 
-  # Evaluate
-  eval(as.call(call), as.environment(object$data))
+  #----- Export everything
+
+  # Also put the data as in the fitted cirls
+  fit$model <- object[["model"]]
+  fit$y <- object[["y"]]
+  fit$x <- object[["x"]]
+
+  # Final list
+  structure(c(fit,
+    list(call = as.call(call), formula = object$formula, terms = mt,
+      data = object$data, offset = offset,
+      control = do.call("glm.control", control),
+      method = "glm.fit", contrasts = attr(x, "contrasts"),
+      xlevels = .getXlevels(mt, mf))),
+    class = c(fit$class, c("glm", "lm")))
 }